Deduplicate get_gcloud_token()
