+++ skills/rhdh-test-plan-review/scripts/gcloud_token.py
@@ -1,0 +1,21 @@
+"""Shared gcloud access token helper for rhdh-test-plan-review scripts."""
+
+import shutil
+import subprocess
+
+
+def get_gcloud_token():
+    """Return (access_token_or_None, error_message_or_None) from gcloud CLI."""
+    gcloud = shutil.which("gcloud")
+    if not gcloud:
+        return None, "gcloud not on PATH — install Google Cloud CLI and add gcloud to PATH"
+
+    result = subprocess.run(
+        [gcloud, "auth", "print-access-token"],
+        capture_output=True,
+        text=True,
+    )
+    token = result.stdout.strip()
+    if token:
+        return token, None
+    return None, "No active gcloud account — run: gcloud auth login --enable-gdrive-access"
